+++ sender.go
@@ -8,6 +8,14 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
+type Sender interface {
+	Update(recoveryScheme)
+	Close()
+}
+
+// TODO: should have access to the mixesoutputs. that means that it should know about all mixes that had already sent,
+// 	and attempt resending them if necessary.
+
 type messageSender struct {
 	round int
 
@@ -38,33 +46,7 @@ func (m *messageSender) worker() {
 		groupedOnions := GroupOnionsByHostname(tosend.onions, m.topology)
 
 		for host, onionsByHost := range groupedOnions {
-			mp := GroupOnionsByMixName(onionsByHost, m.topology)
-
-			msgs := &AddMessagesRequest{
-				Round:    uint32(m.round),
-				Messages: make([]*Messages, len(mp)),
-			}
-
-			i := 0
-
-			for mixToSendTo, onions := range mp {
-				msgs.Messages[i] = &Messages{
-					Messages:        onionsToRepeatedByteArrays(onions),
-					PhysicalSender:  m.address,
-					LogicalSender:   tosend.logicalSender,
-					LogicalReceiver: []byte(mixToSendTo),
-				}
-
-				i += 1
-			}
-
-			rq := &rrpc.Request{
-				Args:    msgs,
-				Reply:   &AddMessagesResponse{},
-				Method:  "/resmix.Mix/AddMessages",
-				Uuid:    fmt.Sprintf("%v-to-%v:from-%v", m.address, host, string(tosend.logicalSender)),
-				Context: context.Background(), // TODO: get valid timeout for this context.
-			}
+			rq := m.createRRPCRequest(onionsByHost, tosend, host)
 
 			if err := m.connections[hostname(host)].DirectCall(rq); err != nil {
 				m.log.Errorln("direct call failed: ", err.Error())
@@ -73,7 +55,38 @@ func (m *messageSender) worker() {
 	}
 }
 
-func NewSender(round int, h string, topology *config.Topology, connections map[hostname]rrpc.ClientConn, input <-chan mixoutput) Sender {
+func (m *messageSender) createRRPCRequest(onionsByHost []Onion, tosend mixoutput, host string) *rrpc.Request {
+	mp := GroupOnionsByMixName(onionsByHost, m.topology)
+
+	msgs := &AddMessagesRequest{
+		Round:    uint32(m.round),
+		Messages: make([]*Messages, len(mp)),
+	}
+
+	i := 0
+
+	for mixToSendTo, onions := range mp {
+		msgs.Messages[i] = &Messages{
+			Messages:        onionsToRepeatedByteArrays(onions),
+			PhysicalSender:  m.address,
+			LogicalSender:   tosend.logicalSender,
+			LogicalReceiver: []byte(mixToSendTo),
+		}
+
+		i += 1
+	}
+
+	rq := &rrpc.Request{
+		Args:    msgs,
+		Reply:   &AddMessagesResponse{},
+		Method:  "/resmix.Mix/AddMessages",
+		Uuid:    fmt.Sprintf("%v-to-%v:from-%v", m.address, host, string(tosend.logicalSender)),
+		Context: context.Background(), // TODO: get valid timeout for this context.
+	}
+	return rq
+}
+
+func NewSender(round int, h string, topology *config.Topology, connections map[hostname]rrpc.ClientConn, input <-chan mixoutput) *messageSender {
 	s := &messageSender{
 		round:       round,
 		topology:    topology,
